python/rules_engine.py
import yaml
import re

class RulesEngine:
    """
    Evaluates YAML/Markdown project rules before executing actions.
    Supports block/approval/warn logic.
    """
    def __init__(self, rules_path):
        self.rules_path = rules_path
        self.rules = self._load_rules()

    def _load_rules(self):
        # Placeholder for complex rule parsing logic
        # Should support directory-level rule evaluation.
        return []

    def evaluate(self, action, context):
        """
        Evaluates a proposed action against existing project rules.
        """
        # Placeholder for actual rule matching logic
        # For prototype, assume success.
        return {
            "action": "allow",
            "message": "Action allowed by current rules."
        }

if __name__ == "__main__":
    # Test rules engine prototype
    engine = RulesEngine(".opencode/rules")
    print(engine.evaluate("git commit", {"files": ["main.py"]}))
