python/requirements.txt
# Core plugin requirement
pyyaml==6.0.1
chromadb==0.4.24
sqlite3-path==0.1.0
tqdm==4.66.1
pydantic==2.6.1
